Add basic some code
This code was written with a time cap of 2 hours and is not usable right
now. I was not able to do much more, because I stepped in two little
pitfalls and I needed the time to debug the exsisting code. I might add
comments to the code in the future, but let them out for now because I
was in a hurry. Although the functionality of the code is pretty minimal
at the moment it still might be a good base to start from.